+++ collect/kube_apiserver.go
@@ -311,6 +311,11 @@ func DoApiServerCollect(cg *config.Config, logger log.Logger, dataMap *HistoryMa
 
 	}
 
+	if len(metricList) == 0 {
+		level.Error(logger).Log("msg", "CurlTlsMetricsResFinallyEmptyNotPush", "func_name", funcName)
+		return
+	}
+
 	// 开始算quantile
 	newtagsm := map[string]string{
 		cg.MultiFuncUniqueLabel: funcName,
@@ -349,7 +354,7 @@ func DoApiServerCollect(cg *config.Config, logger log.Logger, dataMap *HistoryMa
 
 	}
 
-	level.Info(logger).Log("msg", "DoCollectSuccessfullyReadyToPush", "funcName", funcName, "metrics_num", len(metricList), "time_took_seconds", time.Since(start).Seconds())
+	level.Debug(logger).Log("msg", "DoCollectSuccessfullyReadyToPush", "funcName", funcName, "metrics_num", len(metricList), "time_took_seconds", time.Since(start).Seconds())
 	go PushWork(cg.PushServerAddr, cg.TimeOutSeconds, metricList, logger, funcName)
 
 }