+++ collect/kube_controller_manager.go
@@ -154,6 +154,10 @@ func DoKubeControllerCollect(cg *config.Config, logger log.Logger, dataMap *Hist
 			metricList = append(metricList, metric)
 		}
 	}
+	if len(metricList) == 0 {
+		level.Error(logger).Log("msg", "CurlTlsMetricsResFinallyEmptyNotPush", "func_name", funcName)
+		return
+	}
 
 	newtagsm := map[string]string{
 		cg.MultiFuncUniqueLabel: funcName,
@@ -173,7 +177,7 @@ func DoKubeControllerCollect(cg *config.Config, logger log.Logger, dataMap *Hist
 		metricList = append(metricList, mm...)
 
 	}
-	level.Info(logger).Log("msg", "DoCollectSuccessfullyReadyToPush", "funcName", funcName, "metrics_num", len(metricList), "time_took_seconds", time.Since(start).Seconds())
+	level.Debug(logger).Log("msg", "DoCollectSuccessfullyReadyToPush", "funcName", funcName, "metrics_num", len(metricList), "time_took_seconds", time.Since(start).Seconds())
 	go PushWork(cg.PushServerAddr, cg.TimeOutSeconds, metricList, logger, funcName)
 
 }