+++ collect/kube_coredns.go
@@ -127,6 +127,11 @@ func DoKubeCoreDnsCollect(cg *config.Config, logger log.Logger, dataMap *History
 		}
 	}
 
+	if len(metricList) == 0 {
+		level.Error(logger).Log("msg", "CurlTlsMetricsResFinallyEmptyNotPush", "func_name", funcName)
+		return
+	}
+
 	newtagsm := map[string]string{
 		cg.MultiFuncUniqueLabel: funcName,
 	}
@@ -144,7 +149,7 @@ func DoKubeCoreDnsCollect(cg *config.Config, logger log.Logger, dataMap *History
 		metricList = append(metricList, mm...)
 
 	}
-	level.Info(logger).Log("msg", "DoCollectSuccessfullyReadyToPush", "funcName", funcName, "metrics_num", len(metricList), "time_took_seconds", time.Since(start).Seconds())
+	level.Debug(logger).Log("msg", "DoCollectSuccessfullyReadyToPush", "funcName", funcName, "metrics_num", len(metricList), "time_took_seconds", time.Since(start).Seconds())
 	go PushWork(cg.PushServerAddr, cg.TimeOutSeconds, metricList, logger, funcName)
 
 }